move parameterized test to own class
